refactor: split character routes into modular files
