+++ server/routes/characters/occupations.js
@@ -1,0 +1,44 @@
+const ObjectId = require('mongodb').ObjectId;
+const express = require('express');
+const authenticateToken = require('../../middleware/auth');
+
+module.exports = (router) => {
+  const characterRouter = express.Router();
+
+  // Apply authentication to all character routes
+  characterRouter.use(authenticateToken);
+
+  // This section will get a list of all the occupations.
+  characterRouter.route('/occupations').get(async (req, res, next) => {
+    try {
+      const db_connect = req.db;
+      const result = await db_connect
+        .collection('Occupations')
+        .find({})
+        .toArray();
+      res.json(result);
+    } catch (err) {
+      next(err);
+    }
+  });
+
+  // This section will update occupations.
+  characterRouter.route('/update-occupations/:id').put(async (req, res, next) => {
+    const id = { _id: ObjectId(req.params.id) };
+    const db_connect = req.db;
+
+    try {
+      await db_connect.collection('Characters').updateOne(id, {
+        $set: { occupation: req.body },
+      });
+      console.log('Character occupations updated');
+      res.send('User updated successfully');
+    } catch (error) {
+      console.error(error);
+      res.status(500).send('Server error');
+    }
+  });
+
+  router.use(characterRouter);
+};
+
