+++ server/routes/characters/stats.js
@@ -1,0 +1,35 @@
+const ObjectId = require('mongodb').ObjectId;
+const express = require('express');
+const authenticateToken = require('../../middleware/auth');
+
+module.exports = (router) => {
+  const characterRouter = express.Router();
+
+  // Apply authentication to all character routes
+  characterRouter.use(authenticateToken);
+
+  // This section will update stats.
+  characterRouter.route('/update-stats/:id').put(async (req, res, next) => {
+    const id = { _id: ObjectId(req.params.id) };
+    const db_connect = req.db;
+    try {
+      await db_connect.collection('Characters').updateOne(id, {
+        $set: {
+          str: req.body.str,
+          dex: req.body.dex,
+          con: req.body.con,
+          int: req.body.int,
+          wis: req.body.wis,
+          cha: req.body.cha,
+        },
+      });
+      console.log('character stats updated');
+      res.send('user updated sucessfully');
+    } catch (err) {
+      next(err);
+    }
+  });
+
+  router.use(characterRouter);
+};
+
