+++ server/routes/index.js
@@ -6,7 +6,10 @@ require('dotenv').config();
 const auth = require('./auth');
 const users = require('./users');
 const campaigns = require('./campaigns');
-const characters = require('./characters');
+const characterBase = require('./characters/base');
+const characterOccupations = require('./characters/occupations');
+const characterStats = require('./characters/stats');
+const characterHealth = require('./characters/health');
 const skills = require('./skills');
 const feats = require('./feats');
 const equipment = require('./equipment');
@@ -23,7 +26,10 @@ routes.use(async (req, res, next) => {
 auth(routes);
 users(routes);
 campaigns(routes);
-characters(routes);
+characterBase(routes);
+characterOccupations(routes);
+characterStats(routes);
+characterHealth(routes);
 skills(routes);
 feats(routes);
 equipment(routes);
